chore(schema):  update base event view for incremental re-computation (#1205)

+++ src/analytics/private/sqls/redshift/dashboard/clickstream_acquisition_day_user_acquisition.sql
@@ -7,8 +7,8 @@ CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.{{viewName}} (
     new_user_count bigint,
     session_count bigint,
     engagement_session_count bigint,
-    engagement_rate decimal,
-    total_user_engagement_time_minutes bigint,
+    engagement_rate double precision,
+    total_user_engagement_time_minutes double precision,
     avg_user_engagement_time_minutes double precision,
     event_count bigint,
     update_timestamp timestamp DEFAULT getdate()
